Title: 최솟값 3개, Time: 6ms, Memory: 0MB, Status: Wrong Answer - Codetree

+++ 250414/최솟값 3개/top-3-smallest-number.cpp
@@ -22,7 +22,7 @@ int main() {
             int b = -pq.top(); pq.pop();
             int c = -pq.top(); pq.pop();
 
-            int product = a * b * c;
+            long long product = a * b * c;
             cout << product << '\n';
 
             pq.push(-a);
